+++ Demo/Submit-Gist.test.ps1
@@ -1,11 +1,12 @@
-
 <#
 .Synopsis
 	Submit-Gist.ps1 tests.
 
 .Description
-	Task SubmitNotChanged is hardcoded for gist-1303971 C:\ROM\APS\Markdown.tasks.ps1
+	Task SubmitNotChanged is hardcoded for gist-1303971 ...\Markdown.tasks.ps1
 #>
+
+$SubmitNotChangedFile = "C:\-\zip\DEV\MarkdownDeep\Markdown.tasks.ps1"
 
 task MissingFile {
 	$ErrorActionPreference = 'Continue'
@@ -32,7 +33,7 @@ task SubmitNotChanged {
 
 	### 1.
 	Write-Build Cyan "submit - yet no repo, do not keep"
-	$res2 = Submit-Gist.ps1 C:\ROM\APS\Markdown.tasks.ps1 | .{process{$_.Trim()}}
+	$res2 = Submit-Gist.ps1 $SubmitNotChangedFile | .{process{$_.Trim()}}
 
 	# repo must be removed
 	assert (![IO.Directory]::Exists($repo1303971))
@@ -46,7 +47,7 @@ task SubmitNotChanged {
 	### 2.
 	Write-Build Cyan "submit - yet no repo, keep"
 	$res1.text = ''
-	$res2 = Submit-Gist.ps1 C:\ROM\APS\Markdown.tasks.ps1 -Keep
+	$res2 = Submit-Gist.ps1 $SubmitNotChangedFile -Keep
 
 	# repo must exist
 	assert ([IO.Directory]::Exists($repo1303971))
@@ -60,7 +61,7 @@ task SubmitNotChanged {
 	### 3.
 	Write-Build Cyan "submit - repo exists, do not keep"
 	$res1.text = ''
-	$res2 = Submit-Gist.ps1 C:\ROM\APS\Markdown.tasks.ps1
+	$res2 = Submit-Gist.ps1 $SubmitNotChangedFile
 
 	# repo still must exist
 	assert ([IO.Directory]::Exists($repo1303971))